+++ src/main/java/com/example/examplemod/data/ModDataComponents.java
@@ -3,10 +3,13 @@ package com.example.examplemod.data;
 import java.util.function.UnaryOperator;
 
 import com.example.examplemod.ExampleMod;
+import com.mojang.serialization.Codec;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.component.DataComponentType;
 import net.minecraft.core.registries.Registries;
+import net.minecraft.network.codec.ByteBufCodecs;
+import net.minecraft.world.level.block.Block;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.neoforge.registries.DeferredHolder;
 import net.neoforged.neoforge.registries.DeferredRegister;
@@ -17,6 +20,10 @@ public class ModDataComponents {
 
     public static final DeferredHolder<DataComponentType<?>, DataComponentType<BlockPos>> CORDINATES = register(
             "cordinates", builder -> builder.persistent(BlockPos.CODEC).networkSynchronized(BlockPos.STREAM_CODEC));
+
+    public static final DeferredHolder<DataComponentType<?>, DataComponentType<String>> TRESURE_BLOCK_NAME = register(
+            "treasure_block_name",
+            builder -> builder.persistent(Codec.STRING).networkSynchronized(ByteBufCodecs.STRING_UTF8));
 
     public static <T> DeferredHolder<DataComponentType<?>, DataComponentType<T>> register(String name,
             UnaryOperator<DataComponentType.Builder<T>> buildOperator) {